# frontend/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ANALYTE",
    "description": "Bazimi determinist dhe verifikimi i shpjegimeve të analizave laboratorike.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/2fa": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Status",
        "operationId": "status_auth_2fa_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwoFactorStatus"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/2fa/enroll": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Enroll",
        "description": "Nis regjistrimin: një sekret i ri, i dukshëm vetëm në këtë përgjigje. Thirrja e dytë para konfirmimit e\nzëvendëson sekretin; pasi hapi i dytë aktivizohet, kjo pikë kthen 409 dhe sekreti nuk lexohet më.",
        "operationId": "enroll_auth_2fa_enroll_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwoFactorEnrollment"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/2fa/confirm": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Confirm",
        "description": "Aktivizon hapin e dytë me kodin e parë të vlefshëm dhe kthen kodet e rimëkëmbjes, një herë.",
        "operationId": "confirm_auth_2fa_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/2fa/disable": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Disable",
        "description": "Çaktivizon hapin e dytë. Kërkon fjalëkalimin e tanishëm dhe një kod (aplikacioni ose rimëkëmbjeje): një token\naksesi i vjedhur vetëm nuk mjafton. Gabimet kufizohen si te hyrja.",
        "operationId": "disable_auth_2fa_disable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableTwoFactorIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/login/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify Login",
        "description": "Hapi i dytë i hyrjes: sfida nga `login` dhe një kod. Hap seancën kur kodi vlen.",
        "operationId": "verify_login_auth_login_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginVerifyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tokens"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CodeIn": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "CodeIn"
      },
...
      "DisableTwoFactorIn": {
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          },
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "password",
          "code"
        ],
        "title": "DisableTwoFactorIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginVerifyIn": {
        "properties": {
          "challenge": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 20,
            "title": "Challenge"
          },
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "challenge",
          "code"
        ],
        "title": "LoginVerifyIn"
      },
...
      "RecoveryCodes": {
        "properties": {
          "recovery_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recovery Codes"
          }
        },
        "type": "object",
        "required": [
          "recovery_codes"
        ],
        "title": "RecoveryCodes",
        "description": "Kodet e rimëkëmbjes. Shfaqen vetëm një herë: shërbimi mban vetëm HMAC-in e tyre."
      },
...
      "Tokens": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "Tokens"
      },
...
      "TwoFactorEnrollment": {
        "properties": {
          "secret": {
            "type": "string",
            "title": "Secret"
          },
          "otpauth_uri": {
            "type": "string",
            "title": "Otpauth Uri"
          }
        },
        "type": "object",
        "required": [
          "secret",
          "otpauth_uri"
        ],
        "title": "TwoFactorEnrollment",
        "description": "Sekreti dhe URI-ja e aplikacionit. Shfaqen vetëm këtu, një herë; pasi regjistrimi të konfirmohet nuk lexohen më."
      },
      "TwoFactorStatus": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "recovery_codes_remaining": {
            "type": "integer",
            "title": "Recovery Codes Remaining"
          }
        },
        "type": "object",
        "required": [
          "enabled",
          "recovery_codes_remaining"
        ],
        "title": "TwoFactorStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}